Inital commit of all models for phase 2 of the CFMS Update Project.
Included all tables from CFMS data dictionary. Currently, this webapp uses flask restplus to handle all of its response marshalling. This commit will not make sense until the resources commit happens later. There is a basic bootstrap file that is included with this commit that contains basic database inserts for testing purposes.

Future commits to this branch will see resources added for webapp endpoints that work directly with the models created in this commit.

Currently the only issues with this commit is the model base class has inheritance issues that need to be investigated. Through testing, it was found that running a migrate of the database without dropping the previous version of the database would cause a cascading error, with the inability to bootstrap new data to cockroach db. To ensure that this webapp can be tested after adding new data to the models, drop the queue_management database, re-create the queue_management database, provide the proper grants to qsystem, and then init, migrate, update, bootstrap and finally runserver for localhost testing.