Added API calls for generic citizen invite
To invite a generic citizen, the call looks at the quick transaction status of the currently signed in CSR, and tries to find a matching citizen in the queue to invite. If no matching status citizen can be found, then instead we look for any citizen waiting to be queued.