Added feature flagging for the Exams feature that was working on in sprint 2.
Created a field on the office table called exams_enabled_ind. Items effected by this flagging are the hamburger icon, the exams table and the add exam button. If a user somehow navigates to the exams page without using the hamburger icon navigation, and their office is not enabled to view exams, they are greeted with a Coming Soon message.