Individual and Group Exam Notifications
This PR includes exam and exam type model updates for exam returned ind, exam returned tracking number and group exam ind. Back end modifications include two new objects included with the CSRS ME get endpoint that show how many group and individual exams that exist against specific filtering criteria. The front end modifications include functionality modified with the vue action getUsers that looks at the new json objects that are included with other csrs me data. The notification that is presented to the user upon login is tailored to be specific on what type of exams exist based on whats included with the csrs me data. The notification is set to be available for 999 seconds, is dismissible and is only present to the user upon login and at no other time.