Add Invigilator to Booking Modal
In order to book an exam, the ability to include an invigilator must be cinluded on the calendar booking. To do this, the invigilator FK on the exam model needed to be removed and then added to the booking model. Subsequent changes to these models cascaded through end-points in the bookings feature. Once these changes were compelted and the app could compile and run without error, front end changes were made to the bookings modal to include a dropdown option that lists all invigilators available to an office. Afterwards functionality was added to the modal which allowed a CSRto review booking data before submitting.