Make Updates to an Existing Exam Record
CSRs require the ability to update existing exams records in the exam inventory table. This comes in the form of two actions listed as dropdown items in the exam inventory table now. The first action (Edit Exam) shows the user basic information that is displayed in the table, and allows the user to edit said data, then submit the data to the database. The second action (Return Exam) allows the user to enter or edit information with respect to the exam being returned (an ind field and tracking information). Both actions upon submit display a relevent success or failure message depending on the return method response.