Fix issue with Alembic migration history
A revision was added in error further down the revision tree than upstream database revisions. This caused the upstream databases to not pick up a required change when deploying the app. This change correctly inserts the alembic migration file at the top of the chain, thus it should get deployed properly.