Exam Returned Ind Default Value
During testing, it was found out that the exam_returned_ind field on the exam table was set to nullable=false. When a new exam was created, this field was not attached to the POST method, which ended up in a 422 error code. This was fixed by adding a default value of zero to the column so that the nullable property of the field was not violated, which means on creation, exams are not returned.