Add Offsite Location Field
During development of the capture group exam component of the exams feature, it was found out that the exams model/schema required an offsite location field to accomodate offices in the greater vancouver area. This field was added to both the exams model and schema, and tested in postman to make sure that the exams GET endpoint worked.