Code Clean-up: Redundant Vue Action
During end of sprint code clean-up, it was found out that the examsOnLogin action in the index.js file was actually not used, and therefore redundant. This was cleaned up, and the bookings features tested on the front end to ensure that no negative side effects were created post action removal.