Sprint 4 - Swagger/Postman Updates
During the development of the bookings app in sprint 4, a field was added to the exams model/schema, and a GET method was added to list all offices in theq. These changes needed to be reflected in the bookings postman tests, as well as in the swagger yaml file. The postman tests were written and tested before submission of this PR, and were verified to be working as expected.