Fixed various issues with add-group-exam workflow+filtering on exam table
Improved the add exam workflow including addition of flatpickr.js for time selection, an 'add another exam' button at the end, and sepperated the access to the form to another button visible only to the LIAISON role.  Added filters to the exam inventory table for Group vs. Individual, Expired vs Current, and Booked vs Unbooked, as well as modified the location column to handle the longer location data associated with group exam bookings.  also fixed an issue causing the agenda and bookings views to render incorrectly when there are no bookings to display