Monthly Financial Reports - Backend Changes
As a finanace member, a bookings user will require to access an ability to gerenate a report or csv file to help generate a report on monthly/yearly revenues incurred from exams. In order to do this, a new end-point was added to GET all exams based on two url parameteres (start_date and end_date). Based on these parameters, the GET looks at all exams that were booked between those dates (exam and booking model are joined, then queried). Along with this new end-point, this PR includes a new set of postman tests to ensure that this end-point returns exactly what is required in order to generate these reports.