Booking Foreign Key Changes
During a live client demo, it was determined that there needed to be changes made to the booking model and schema in order to accomplish the creation and listing of group exams in the booking feature. This PR addresses these requirements by making the room_id foreign key nullable, and adding an office_id foreign key that references the offices table. One additional PR will be required to update postman tests and swagger yaml files to communicate these changes to the client.