Alembic Migrate File Changes
During development of the bookings feature, a field offsite_location field was added to the exam table to accomodate group exam bookings. It was initially alotted a size of 50 chars, however it was found out that this field should've been alotted a size of 255. When this change was made to the exam model, alembic didn't pick up the change, and therefore this change had to be included on the next migration file that alembic generated.