Bookings Backend Changes - Postman/Swagger Updates
In order to accomodate changing requirements for group exams to be booked, changes to the booking model were needed. This included adding a office_id foreign key, and making the room_id foreign key nullable on the booking model/schema. The returned booking object also was required to contain an inflated office object as well when the bookings GET method was called. Once these changes were made, updates to the swagger documention and postman tests were also required. This PR contains newly updated postman tests for GET, POST, PUT and DEL methods, as well as updated fields on the swagger documentation.