Modify End Point to Return CSV
According to requirements provided to the developers, the ability for a CSR/GA/Liaison to generate a financial report entailed the generation of a CSV file in order to build this report. This report includes fields from the exam, booking, room and offices tables. This file should then be exported directly to the clients browser for them to use. The modifications to the GET request that is used to generate the csv file were tested in postman. Future modifications to this would possibly include the office name or timestamp in the file name of the export csv file.