Monthly Financial Reporting - Office/Exam Group Filtering
A CSR/GA/Liaison will require the ability to generate financial reports based on type of exam as well as for a specific office. This PR contains the added functionality to add an exam_type and office_name url parameter to the url used to export a csv to the clients browser. At this moment in time, both parameters are not required or used when creatiung the CSV, but will be required in future development when generating these csv files.