Booking Exam From Exam Admin Inventory Table
After client testing, it was determined that the exam inventory table on the exam admin feature required the ability for exams that were not already booked, to be booked. Logic in the PR looks at an exam, and if it has not been booked, a dropdown item in the actions field will appear to Book this exam. Once clicked, the user is taken to the booking calendar feature (week view) and placed into scheduling mode with the exam that was clicked to be added. There the user can click on the calendar to place the exam into whatever time slot, in whatever room they want.