Sprint 5 (#170)
* Monthly Financial Reports - Backend Changes

As a finanace member, a bookings user will require to access an ability to gerenate a report or csv file to help generate a report on monthly/yearly revenues incurred from exams. In order to do this, a new end-point was added to GET all exams based on two url parameteres (start_date and end_date). Based on these parameters, the GET looks at all exams that were booked between those dates (exam and booking model are joined, then queried). Along with this new end-point, this PR includes a new set of postman tests to ensure that this end-point returns exactly what is required in order to generate these reports.

* Modify End Point to Return CSV

According to requirements provided to the developers, the ability for a CSR/GA/Liaison to generate a financial report entailed the generation of a CSV file in order to build this report. This report includes fields from the exam, booking, room and offices tables. This file should then be exported directly to the clients browser for them to use. The modifications to the GET request that is used to generate the csv file were tested in postman. Future modifications to this would possibly include the office name or timestamp in the file name of the export csv file.

* Addressed issues stemming from Jan 25th testing session re: Group Exams

Covered the issues detailed in 5 of the trello cards generated after this session:
-Changed the behaviour of the select exam type dropdown in Step 1 of Add Exam (both types) - you can now click anywhere on the input to expand the menu
-Filtered the data fed into the above noted dropdown - only shows group exams when adding group and individual when adding individual
-Added logic to create a virtual room resource in the calendar for group exams called offsite.  now rendering all group exams to display into it after they have been initially entered by LIAISONs
-Modified the calendar to hide and show the offsite room as appropriate (hidden when booking is happening, visible otherwise)
-Fixed the 'something went wrong' which displays when an exam post fails-now possible to try again or review captured details.
-Fixed the display of the exam inventory table when rendering inside the modal.  contents now 'fit' in the modal.

* Fixed issue with modal padding

Created class in q.css with zero padding and applied to necessary modals via body-class prop as it should have been done.

* Booking Foreign Key Changes

During a live client demo, it was determined that there needed to be changes made to the booking model and schema in order to accomplish the creation and listing of group exams in the booking feature. This PR addresses these requirements by making the room_id foreign key nullable, and adding an office_id foreign key that references the offices table. One additional PR will be required to update postman tests and swagger yaml files to communicate these changes to the client.

* Alembic Migrate File Changes

During development of the bookings feature, a field offsite_location field was added to the exam table to accomodate group exam bookings. It was initially alotted a size of 50 chars, however it was found out that this field should've been alotted a size of 255. When this change was made to the exam model, alembic didn't pick up the change, and therefore this change had to be included on the next migration file that alembic generated.

* Bookings Backend Changes - Postman/Swagger Updates

In order to accomodate changing requirements for group exams to be booked, changes to the booking model were needed. This included adding a office_id foreign key, and making the room_id foreign key nullable on the booking model/schema. The returned booking object also was required to contain an inflated office object as well when the bookings GET method was called. Once these changes were made, updates to the swagger documention and postman tests were also required. This PR contains newly updated postman tests for GET, POST, PUT and DEL methods, as well as updated fields on the swagger documentation.

* Updated bookings resources to work with new data model

* Monthly Financial Reporting - Office/Exam Group Filtering

A CSR/GA/Liaison will require the ability to generate financial reports based on type of exam as well as for a specific office. This PR contains the added functionality to add an exam_type and office_name url parameter to the url used to export a csv to the clients browser. At this moment in time, both parameters are not required or used when creatiung the CSV, but will be required in future development when generating these csv files.

* Fixed date/time input components and jitter on navigation

Changed the datetime inputs used in the add-exam form (both types) and fixed the naviagation jitter.

* Booking Exam From Exam Admin Inventory Table

After client testing, it was determined that the exam inventory table on the exam admin feature required the ability for exams that were not already booked, to be booked. Logic in the PR looks at an exam, and if it has not been booked, a dropdown item in the actions field will appear to Book this exam. Once clicked, the user is taken to the booking calendar feature (week view) and placed into scheduling mode with the exam that was clicked to be added. There the user can click on the calendar to place the exam into whatever time slot, in whatever room they want.