Admin Ability to Flag CSR as ITA Designate
For future work regarding how the app will notify CSRs of exams that have been received, there needed to be a flag attached to the csr model called 'ita_designate'. This flag was to be modelled after the receptionist_ind field on the csr model, and had to be editable through theq admin interface. This PR contains the model/schema updates for this field as well as the front end changes to the admin interface that allow staff with the ability to see the admin interface to edit this field.