Remove Ability to Drag For Events on Bookings Calendar
During demo testing, it was found that the bookings calendar was selectable by default, allowing users to drag events onto the calendar without having an event in queue to be placed on the calendar. This was not required functionality on the calendar, and was deemed a bug. This bug was fixed by how the calendar.vue file looks for URL parameters, and how it sets the selectable option to true or false.