Scheduling Appointments - API End Points
In order for the front end to use the model/schema for appointments, a GET detail, GET list, POST, PUT and DELETE http method needed to be developed. These end-points follow standards for office ID checking throughout the rest of the bookings feature. These end-points were tested using postman to ensure that proper responses were made to the actions taken against the API.