Created new edit exams modal and form
Created a new edit exam form that has multiple setups with different fields for the various types of CSRs that allows editing of all the required fields.  Built a new office search functionality that allows for typing directly into the input field with suggestions populating beneath or inputing the office number.
Also fixed previously undocumented bug which caused a exam_received_date to be set to the current date during the initial capture of a group exam--value now remains unset for branch staff to update when appropriate.