Appointments Version File
During development it was found that UTC datetime objects were causing an issue in the backend. To mitigate these issues, the version file that was created for the appointments table removed the UTC datetime contraints on 3 specific columns, and replaced them with basic SQL Alchemy datetime objects. To re-inforce the UTC contstraint, the models will maintain the UTC datetime datatype.