Generate Financial Report - Exam Type Filtering
Further development on the generate CSV modal required exam type filtering to be completed. Based on the exam type elected via radio button, a url parameter is passed to the GET method used to generate the final csv. Based on the parameter provided, dynamic filtering is done to the exam object created via query.