Fixed issues when click 'reschedule' and 'schedule' from exam inventory
Scheduling and rescheduling exams from the exam inventory view now work properly and display of the offsite room is now as expected.  Simplified the calendar component, removing numerous redundant state properties and mutations, as well as removed the use of the router for any of the calendar's functionality.  Removed several other unused state properties and related mutations, fixed a bug in the add individual ita exam workflow where if logging two exams in a row, for the second exam the "Exam Received Today" option was not preselected.  Also fixed the bug causing jitter when opening the hamburger menu in some circumstances.