Fixed issue with exam inventory modal filters
Fixed issue which was causing exam inventory filters not to work correctly following the addition of the sbc_staff_invigilated flag to the data model.  also fixed an issue with the tabular data display in the edit-group-exam-booking-modal and a bug with the inivigilator dropdown in same.  also added the ability to PUT an 'unassigned' invigilator to a group exam.