Sprint 6 (#179)
* Scheduling Appointments - Models and Schemas

CSRs were determined to need the ability to schedule smaller appointments on the bookings calendar feature for reasons other than exams. For this, an appointments model and schema were developed such that a separate calendar feature can be developed, separate from the exam bookings calendar. Future development will include GET, POST, PUT and DELETE http methods, as well as updates to postman tests and swagger api spec.

* Admin Ability to Flag CSR as ITA Designate

For future work regarding how the app will notify CSRs of exams that have been received, there needed to be a flag attached to the csr model called 'ita_designate'. This flag was to be modelled after the receptionist_ind field on the csr model, and had to be editable through theq admin interface. This PR contains the model/schema updates for this field as well as the front end changes to the admin interface that allow staff with the ability to see the admin interface to edit this field.

* Remove Ability to Drag For Events on Bookings Calendar

During demo testing, it was found that the bookings calendar was selectable by default, allowing users to drag events onto the calendar without having an event in queue to be placed on the calendar. This was not required functionality on the calendar, and was deemed a bug. This bug was fixed by how the calendar.vue file looks for URL parameters, and how it sets the selectable option to true or false.

* Scheduling Appointments - API End Points

In order for the front end to use the model/schema for appointments, a GET detail, GET list, POST, PUT and DELETE http method needed to be developed. These end-points follow standards for office ID checking throughout the rest of the bookings feature. These end-points were tested using postman to ensure that proper responses were made to the actions taken against the API.

* Appointments Version File

During development it was found that UTC datetime objects were causing an issue in the backend. To mitigate these issues, the version file that was created for the appointments table removed the UTC datetime contraints on 3 specific columns, and replaced them with basic SQL Alchemy datetime objects. To re-inforce the UTC contstraint, the models will maintain the UTC datetime datatype.

* Appointments Postman Tests

In order for the client to maintain the bookings feature, postman tests need to be created and maintained for the new appointments models/schema. All tests have be run against test data and pass with 100% completion.

* Exam Model Backend Changes

Scott required some modifications to the exam model post refactor. This included dropping the exam_received ind and making the expiry_date field nullable. Bootstrap data was modified to meet this change request.

* Created new edit exams modal and form

Created a new edit exam form that has multiple setups with different fields for the various types of CSRs that allows editing of all the required fields.  Built a new office search functionality that allows for typing directly into the input field with suggestions populating beneath or inputing the office number.
Also fixed previously undocumented bug which caused a exam_received_date to be set to the current date during the initial capture of a group exam--value now remains unset for branch staff to update when appropriate.

* Generate CSV Date Picker

After testing, it was determined that the generate CSV modal's datepicker was not using the same datepicker that the rest of Bookings was using. This was changed by importing the DatePicker component from the vu2-datepicker library and replacing the old date picker with the new one.

* Generate Financial Report - Exam Type Filtering

Further development on the generate CSV modal required exam type filtering to be completed. Based on the exam type elected via radio button, a url parameter is passed to the GET method used to generate the final csv. Based on the parameter provided, dynamic filtering is done to the exam object created via query.

* Exam Model Postman Updates

Previously in sprint-6, modifications were made to the exam model, which required modifications to be made to the postman tests. All modifications have been tetsted to ensure that all test cases pass 100%
.

* Fixed issues when click 'reschedule' and 'schedule' from exam inventory

Scheduling and rescheduling exams from the exam inventory view now work properly and display of the offsite room is now as expected.  Simplified the calendar component, removing numerous redundant state properties and mutations, as well as removed the use of the router for any of the calendar's functionality.  Removed several other unused state properties and related mutations, fixed a bug in the add individual ita exam workflow where if logging two exams in a row, for the second exam the "Exam Received Today" option was not preselected.  Also fixed the bug causing jitter when opening the hamburger menu in some circumstances.

* Improved look of booking modal/added default invigilator

The invigilator field now looks to the exam type and defaults to SBC staff when appropriate

* Fixed issue with highest order element being slot instead of template

* Fixed issue with exam inventory modal filters

Fixed issue which was causing exam inventory filters not to work correctly following the addition of the sbc_staff_invigilated flag to the data model.  also fixed an issue with the tabular data display in the edit-group-exam-booking-modal and a bug with the inivigilator dropdown in same.  also added the ability to PUT an 'unassigned' invigilator to a group exam.

* Fixed issue with hamburger menu jitter in ie

fixed the issue of navigation jitter when opening the hamburger menu affecting internet explorer