Fixed issue with exam inventory "Scheduled" column
Previously, the scheduled column would show the unscheduled icon (yellow warning sign) for exams that were actually scheduled.  This was due to the recent addition of the 'sbc_staff_invigilated' field to the data model.  Added logic to check this field as well as invigilator_id in determining whether or not an exam has been scheduled.