Group Exam Reporting
During feature testing, it was found that group exams were not being reported at all. This was due to the room table not being left joined onto the booking table, which was then being joined onto the exam table. The room table being joined normally doesn't work for group exams, as they do not have a room, they are held in an offsite location. Once this was changed to a left join, this issue was resolved. Future work will be to rework line 117 of the exam_export_list file to look for the group_exam_ind == 1 once group exam creation is resolved.