Return Exam Functionality
During testing it was found that the return exam functionality in the exam inventory table was broken. The process to break this function was to create an exam, book the exam and then return the exam immediately after booking. This was fixed by removing the selectedExam state from the return exam form modal and access the fields object instead.