Delete Exam Functionality
During testing, it was found out that the bookings feature lacked a function to delete an exam, not just the exam's booking. The exam-inventory table was modified to have a 'delete exam' dropdown action, which opens a confirmation modal. This modal displays basic exam information to the user before the user deletes the exam. This function has been tested using the CSR, GA and LIAISON roles.