Create Column Pesticide Designate/Nullable Event ID
To meet client requirements for how GAs handle Pesticide Exams, the csr model required a new field, 'pesticide_designate'. This column will allow for the pesticide option in the non-ita exam create modal to be broken away, to make a new crete exam option only for pesticide exams. This PR also contains an update to the exam model, making event ID nullable. This change was made due to the timing of Scott Rumsby working on something that required this change while I was making the required changes to the CSR model.