Completed changes to make option event_id and fixed AddExamModal
Event_id is now optional when posting an exam to the backend, and for ITA Exams, a minimum length of 5 digits is enforced by the frontend.  The summary page now displas correctly for all exam types, and the expiry date field is prepopulated with a date: 90 days in tbe future for individual ITA exams and 60 days in the future for Other Exam types.  The exam_method field prepopulates to 'paper'.  Also, fixed the bug causing it to not be possible to submit an 'other' exam type and the reset modal bu
g.