Mender Integration for RPI (client) and OpenShift (server) (#184)
* Adds templates for running Mender application in OpenShift

This commit provides multiple templates for launching the Mender application in OpenShift. This includes BuildConfigs for generating all of the ImageStreams necessary for running the application.

Please be advised: some of the BuildConfigs modify the underlying Docker Images which ship from Mender. Most of these modifications are to modify directory permissions to allow non-root users. However, some of the modifications do alter configurations of the underlying micro-applications (e.g. `mender-api-gateway`). In instances where the underlying micro-application configuration has been modified, great care has been taken to NOT modify applciation code so as to not mess with future updates, **however** some of the modifications use the `sed` and will not necessarilly fail if upstream files change. **THEREFORE** please take great care when upgrading to newer releases of the underlying Docker Images.

In order to increase flexibility within the configration, the application has been split into multiple templates along logical boundaries within the application, specifically:

* mender-mongodb (mender-mongodb)
* mender-conductor (mender-conductor, mender-redis, mender-elasticsearch)
* mender-components (mender-useradm, mender-deployments, mender-user-auth, mender-inventory)
* mender-frontend (mender-api-gateway, mender-gui)

Each of these templates also uses a label--`subapp`--to define those groups of objects. This makes deleting and scaling (with the `-l subapp=<subapp name> flag) easier for these groups.

The `./mender/openshift/mender_deployment_quickstart.md` file provides the quickest way to launch the application.

---
Additional Information:
* None of the DeploymentConfigs have been tuned (CPU/Memory usage). This should be done as the project is put under load.
* Some of the health monitoring for DeploymentConfigs have been commented out. They were causing failures. This should eventually be dealt with...

* small edit to conductor template. removed ENV which is unncessary.

* Updated Smartboard to support features used for caching and loading in the Raspberry Pi client

* Adds top-level Mender README and updates README in ./image-tools

The top-level README give a basic overview of the whole Mender section of this repo

The README in ./image-tools basically expands on the previous info and attempts to add more clarity.

* Added all code and tools for Raspbian Smartboard image tools

* small formatting edits to mender/image-tools README

* Edits to image-tools/readme based on handoff

includes:
* fixed command error in image-tools/readme
* added QEMU setup stage in mender/image-tools README
* small edits to mender/image-tools README for clartiy for the 'Menderizing the Image' step
* modified README to explain what to do if mender-convert fails
* added note about VM issues in README
* added notes about chromium blue screen

* modify scripts in image-tools directory to automatically build docker images

* removed automatic docker build steps for bootstrap-builder/generate-image.sh and moved this step to the actual sub scripts

* Fix resolv issues

* small edit of image-tools/readme

* Restored cert checking on watchdog and updater to prevent chromium from starting before the clock gets updated.

* Image generation now works in output folder and is non-destructive for input. Fix typo in arrtifact builder

* added missing component-deployment template needed for full OpenShift deployment

* updated mender quickstart to fix a ENV issue for the mender-component-deployment stage

* Updated naming of bootstrap image for clairity

* Fixed pathing issue for artifact building

* Fix issue where ntp will fight with timesyncd

* adds small change to image-tools README for consistency around all 'bootstrap-builder' commands

* Streamline package removal into 01-dependency step. Tighten package remove, and remove ntp.

* Update README to not include base input/ path

* changed default path of mongodb persistant volume from /data/db to /var/lib/mongodb/data

* updated quickstart with more detail about order of operations around deploymentconfigs

* Remove black border on smartboard video

* Tweak smartboard backup video layout. Ensure sites is reloaded after artifact deploy and reboot