Pesticide Exams - Provide Functionality for Designate
Pesticide exams needed to be broken out of the "Add Non-ITA Exam" button such that a pesticide designate liaison could add these exams on their own. This button that is viewable on the exam inventory page can only be seen by csrs with the field pesticide_designate set to 1. This Add Exam functionality mirrors the functionality of the Non-ITA exam creation, but incorperates a select office chooser as well, and displays a notes field on the summary. This functionality has been tested with the GA role and LIAISON role, both with and without the pesticide_designate column set to 1 to ensure that this functionality is completely on its own.