Added exam_type to the exam inventory table
Also, changed the actions column to only display options 'edit exam' and 'delete exam' when LIAISON is viewing inventory out of their own office.  Also, fixed a bug in the summary/confirmation view of the Add Exam Modal where there was a missing < in the html markup causing display issues.