Added Ability to Schedule Challenger Exam
Created an appropriate questions object for Challenger Exams.  Created a store module for the addExamModal and relocated some of the addExamModals data and logic into this module.  Modified the BookingModal, the SchedulingIndicator and the AddExamModal to support onsite or offsite booking of thechallenger exams.  Also modified the ExamInventoryTable to support the Challenger Exams.