Admin Invigilator/Room Feedback
Client testing found that GA access wasn't fully enabled for both
invigilaor and room panels. Also, clients asked for the return results
for both panels to be different based upon client role code. If the user
role code is GA, the return results for both room and invigilator will
be specific to that users office. If the users role code is SUPPORT,
then that user will see ALL results for ALL offices.