Sprint 8 1 (#194)
* Calendar Room Colors

Clients indicated that they would like the default colors for onsite and offsite rooms to be changed. Manage.py now includes the default hex value for lighter daisy, as well as modifiying the calendar component and store to make sure offsite rooms now show up as lighter coral.

* Appointments - Page and Navigation

As required by clients, the bookings feature needs a separate page for an appointments calendar. To begin development on this, a component for the page and its buttons needed to be created, and additions to the application router and nav bar needed to be made in order to navigate to this new page. Future development for this page will include implementing a calendar component on the page that has similar functionality to the exams calendar, but implements black out periods.

* Admin Invigilator/Room Feedback

Client testing found that GA access wasn't fully enabled for both
invigilaor and room panels. Also, clients asked for the return results
for both panels to be different based upon client role code. If the user
role code is GA, the return results for both room and invigilator will
be specific to that users office. If the users role code is SUPPORT,
then that user will see ALL results for ALL offices.

* Added Ability to Schedule Challenger Exam

Created an appropriate questions object for Challenger Exams.  Created a store module for the addExamModal and relocated some of the addExamModals data and logic into this module.  Modified the BookingModal, the SchedulingIndicator and the AddExamModal to support onsite or offsite booking of thechallenger exams.  Also modified the ExamInventoryTable to support the Challenger Exams.

* Added Offsite visibility toggle to calendar and corrected typo

Changed the way the calendar renders resources to the intended way of asynchronously loading roomResources via resource loader function in config.  Made provisions to automatically show or hide the offsite room depending on user preference (in the store) and persist this preference after naviagion.  added toggle switch.  also fixed some typos in ExamInventoryTable.