Sprint 8 2 (#197)
* Remove Invigilator Assign Later Option

During client testing it was found that the invigilator "assign later"
option in the booking modal needed to be removed for individual exams.
This option exists as long as the Group Exam Ind field is set to 0 for
any given exam.

* Edit Non-ITA Exam Details Exam Types

During testing, it was found that the exam type drop down items for non-ita exams were showing up as the ITA exam types. This was fixed by placing a conditional filter on the examType objects returned based on whether or not the ita_ind field is set to 0 or 1.

* Fixed Bugs Preventing Office Agenda from Displaying Events

Fixed the issues related to certain exams not showing on the Office Agenda, added support for the current exam/booking data models (sbc staff invigilators), and fixed the issue resulting in non-exam events not displaying.  added a user-friendly message to display when there are no exam events in scope and added the ability to click the offsite icon to display the offsite details in the row details.  Also fixed...
-Bug preventing Contract Invigilator being assigned from BookingModal initially
-Bug Preventing Contract Invigilator being assigned for Challenger Exams
-Bug related to pressing enter key on AddExamModal and Calendar view causing refresh
-Bug related to exam filtering on ExamInventoryTable
-Bug related to Yellow Triangle on ExamInventoryTable