Appointments Feature Flagging
During client testing, it was found out that the Appointments Feature needed to be separate from the Exams Feature flagging, due to some (not all) offices using the appointments feature. This PR adds a appointments_enabled_ind field to the office model/schema which controls whether or not the the feature shows up in theq nav menu or not.