Contact Info for Non Group Exam Bookings
During client testing, it was found that when non-group exam and non exam events were booked, that client contact information needed to be collected. This required a new field to be added to the booking model and front end changes to be made to the confirm booking and edit booking modals such that these details could be gathered from clients.