Timezones
During client testing, it was found that due to group exams being created by GA's that might not belong to the office that they're creating the exam form (in the same timezone), that timezones needed to be dealt with during group exam creation. To accomplish this, a timezone model/schema was created and attached to the office model/schema through foreign key. This would allow users to choose one of the three timezones available in BC, and attach it to a specific office through the flask admin panels for theq application. This will allow front end development to easily create a group exam in the timezone of the office that the exam belongs to.