Liaison Designate Re-factor
During review/testing of previous PRs, it was found that the liaison designate getter should return a boolean value, and be renamed to include an 'is'. The getter is now named 'is_liaison_designate' and returns a true or false value. This refactor also includes modifying all files that used this getter previously, modifiying the name, and how it was used. This re-factor has been full tested to ensure that functionality has not been lost.